Extract Years With Observations from an Indicator Map

Takes an "indicator_map" object and determines the years for which observation data exists.

Usage

get_observed_years(x)

Arguments

x

An "indicator_map" object containing calculated indicator values associated with grid cells.

Value

A data frame with two columns:

years: A sequence of years covering the range of observations.

occurrences: A logical vector indicating if observations exist for each year (TRUE if present, FALSE if absent).

Examples

get_observed_years(example_cube_1)
#>     years occurrences
#> 1    1751        TRUE
#> 2    1752       FALSE
#> 3    1753       FALSE
#> 4    1754       FALSE
#> 5    1755       FALSE
#> 6    1756       FALSE
#> 7    1757       FALSE
#> 8    1758       FALSE
#> 9    1759       FALSE
#> 10   1760       FALSE
#> 11   1761       FALSE
#> 12   1762       FALSE
#> 13   1763       FALSE
#> 14   1764       FALSE
#> 15   1765       FALSE
#> 16   1766       FALSE
#> 17   1767       FALSE
#> 18   1768       FALSE
#> 19   1769       FALSE
#> 20   1770       FALSE
#> 21   1771       FALSE
#> 22   1772       FALSE
#> 23   1773       FALSE
#> 24   1774       FALSE
#> 25   1775       FALSE
#> 26   1776       FALSE
#> 27   1777       FALSE
#> 28   1778       FALSE
#> 29   1779       FALSE
#> 30   1780       FALSE
#> 31   1781       FALSE
#> 32   1782       FALSE
#> 33   1783       FALSE
#> 34   1784       FALSE
#> 35   1785       FALSE
#> 36   1786       FALSE
#> 37   1787       FALSE
#> 38   1788       FALSE
#> 39   1789       FALSE
#> 40   1790       FALSE
#> 41   1791       FALSE
#> 42   1792       FALSE
#> 43   1793       FALSE
#> 44   1794       FALSE
#> 45   1795       FALSE
#> 46   1796       FALSE
#> 47   1797       FALSE
#> 48   1798       FALSE
#> 49   1799       FALSE
#> 50   1800       FALSE
#> 51   1801       FALSE
#> 52   1802       FALSE
#> 53   1803       FALSE
#> 54   1804       FALSE
#> 55   1805       FALSE
#> 56   1806       FALSE
#> 57   1807       FALSE
#> 58   1808       FALSE
#> 59   1809       FALSE
#> 60   1810       FALSE
#> 61   1811       FALSE
#> 62   1812       FALSE
#> 63   1813       FALSE
#> 64   1814       FALSE
#> 65   1815       FALSE
#> 66   1816       FALSE
#> 67   1817       FALSE
#> 68   1818       FALSE
#> 69   1819       FALSE
#> 70   1820        TRUE
#> 71   1821       FALSE
#> 72   1822       FALSE
#> 73   1823       FALSE
#> 74   1824       FALSE
#> 75   1825       FALSE
#> 76   1826       FALSE
#> 77   1827       FALSE
#> 78   1828       FALSE
#> 79   1829       FALSE
#> 80   1830       FALSE
#> 81   1831       FALSE
#> 82   1832       FALSE
#> 83   1833       FALSE
#> 84   1834       FALSE
#> 85   1835       FALSE
#> 86   1836       FALSE
#> 87   1837       FALSE
#> 88   1838       FALSE
#> 89   1839       FALSE
#> 90   1840       FALSE
#> 91   1841       FALSE
#> 92   1842       FALSE
#> 93   1843       FALSE
#> 94   1844       FALSE
#> 95   1845       FALSE
#> 96   1846       FALSE
#> 97   1847       FALSE
#> 98   1848       FALSE
#> 99   1849       FALSE
#> 100  1850       FALSE
#> 101  1851       FALSE
#> 102  1852       FALSE
#> 103  1853       FALSE
#> 104  1854       FALSE
#> 105  1855       FALSE
#> 106  1856       FALSE
#> 107  1857       FALSE
#> 108  1858       FALSE
#> 109  1859       FALSE
#> 110  1860       FALSE
#> 111  1861       FALSE
#> 112  1862        TRUE
#> 113  1863        TRUE
#> 114  1864       FALSE
#> 115  1865       FALSE
#> 116  1866       FALSE
#> 117  1867       FALSE
#> 118  1868       FALSE
#> 119  1869       FALSE
#> 120  1870        TRUE
#> 121  1871       FALSE
#> 122  1872        TRUE
#> 123  1873       FALSE
#> 124  1874        TRUE
#> 125  1875       FALSE
#> 126  1876        TRUE
#> 127  1877       FALSE
#> 128  1878       FALSE
#> 129  1879        TRUE
#> 130  1880       FALSE
#> 131  1881        TRUE
#> 132  1882       FALSE
#> 133  1883       FALSE
#> 134  1884        TRUE
#> 135  1885       FALSE
#> 136  1886        TRUE
#> 137  1887       FALSE
#> 138  1888       FALSE
#> 139  1889       FALSE
#> 140  1890       FALSE
#> 141  1891       FALSE
#> 142  1892       FALSE
#> 143  1893        TRUE
#> 144  1894        TRUE
#> 145  1895       FALSE
#> 146  1896        TRUE
#> 147  1897        TRUE
#> 148  1898        TRUE
#> 149  1899       FALSE
#> 150  1900       FALSE
#> 151  1901        TRUE
#> 152  1902        TRUE
#> 153  1903        TRUE
#> 154  1904       FALSE
#> 155  1905        TRUE
#> 156  1906       FALSE
#> 157  1907        TRUE
#> 158  1908        TRUE
#> 159  1909        TRUE
#> 160  1910        TRUE
#> 161  1911        TRUE
#> 162  1912       FALSE
#> 163  1913        TRUE
#> 164  1914        TRUE
#> 165  1915       FALSE
#> 166  1916       FALSE
#> 167  1917        TRUE
#> 168  1918       FALSE
#> 169  1919       FALSE
#> 170  1920        TRUE
#> 171  1921        TRUE
#> 172  1922       FALSE
#> 173  1923        TRUE
#> 174  1924        TRUE
#> 175  1925        TRUE
#> 176  1926        TRUE
#> 177  1927        TRUE
#> 178  1928       FALSE
#> 179  1929        TRUE
#> 180  1930        TRUE
#> 181  1931        TRUE
#> 182  1932        TRUE
#> 183  1933        TRUE
#> 184  1934        TRUE
#> 185  1935        TRUE
#> 186  1936        TRUE
#> 187  1937        TRUE
#> 188  1938        TRUE
#> 189  1939        TRUE
#> 190  1940        TRUE
#> 191  1941        TRUE
#> 192  1942        TRUE
#> 193  1943        TRUE
#> 194  1944        TRUE
#> 195  1945        TRUE
#> 196  1946        TRUE
#> 197  1947        TRUE
#> 198  1948        TRUE
#> 199  1949        TRUE
#> 200  1950        TRUE
#> 201  1951        TRUE
#> 202  1952        TRUE
#> 203  1953        TRUE
#> 204  1954        TRUE
#> 205  1955        TRUE
#> 206  1956        TRUE
#> 207  1957        TRUE
#> 208  1958        TRUE
#> 209  1959        TRUE
#> 210  1960        TRUE
#> 211  1961        TRUE
#> 212  1962        TRUE
#> 213  1963        TRUE
#> 214  1964        TRUE
#> 215  1965        TRUE
#> 216  1966        TRUE
#> 217  1967        TRUE
#> 218  1968        TRUE
#> 219  1969        TRUE
#> 220  1970        TRUE
#> 221  1971        TRUE
#> 222  1972        TRUE
#> 223  1973        TRUE
#> 224  1974        TRUE
#> 225  1975        TRUE
#> 226  1976        TRUE
#> 227  1977        TRUE
#> 228  1978        TRUE
#> 229  1979        TRUE
#> 230  1980        TRUE
#> 231  1981        TRUE
#> 232  1982        TRUE
#> 233  1983        TRUE
#> 234  1984        TRUE
#> 235  1985        TRUE
#> 236  1986        TRUE
#> 237  1987        TRUE
#> 238  1988        TRUE
#> 239  1989        TRUE
#> 240  1990        TRUE
#> 241  1991        TRUE
#> 242  1992        TRUE
#> 243  1993        TRUE
#> 244  1994        TRUE
#> 245  1995        TRUE
#> 246  1996        TRUE
#> 247  1997        TRUE
#> 248  1998        TRUE
#> 249  1999        TRUE
#> 250  2000        TRUE
#> 251  2001        TRUE
#> 252  2002        TRUE
#> 253  2003        TRUE
#> 254  2004        TRUE
#> 255  2005        TRUE
#> 256  2006        TRUE
#> 257  2007        TRUE
#> 258  2008        TRUE
#> 259  2009        TRUE
#> 260  2010        TRUE
#> 261  2011        TRUE
#> 262  2012        TRUE
#> 263  2013        TRUE
#> 264  2014        TRUE
#> 265  2015        TRUE
#> 266  2016        TRUE
#> 267  2017        TRUE
#> 268  2018        TRUE
#> 269  2019        TRUE
#> 270  2020        TRUE
#> 271  2021        TRUE
#> 272  2022        TRUE
#> 273  2023        TRUE