Calculate Mean Year of Occurrence Over Space

This function estimates the relative newness of records in a data cube by calculating the mean year of occurrence over a gridded map. It prepares the data, creates a grid, calculates the indicator value for each grid cell, and formats the output into an S3 object.

Usage

newness_map(data, ...)

Arguments

x

A data cube object (class 'processed_cube').

cell_size

Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)

level

Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')

region

The region of interest (e.g., "Europe"). (Default: "Europe")

cube_crs

The projection of the cube. (Default: "EPSG:3035")

first_year

Exclude data before this year. (Uses all data in the cube by default.)

last_year

Exclude data after this year. (Uses all data in the cube by default.)

Value

An S3 object with the classes 'indicator_map' and 'newness' containing the calculated indicator values and metadata.

Examples

n_map <- newness_map(example_cube_1)
#> Error in compute_indicator_workflow(data, type = "newness", dim_type = "map",     ...): object 'final_year' not found
plot(n_map)
#> Error in eval(expr, envir, enclos): object 'n_map' not found