Calculate Evenness Over Time or Space

Calculate evenness over a gridded map or as a time series (see 'Details' for more information).

Usage

pielou_evenness_map(data, ...)

pielou_evenness_ts(data, ...)

williams_evenness_map(data, ...)

williams_evenness_ts(data, ...)

Arguments

data

A data cube object (class 'processed_cube').

...

Arguments passed on to compute_indicator_workflow

ci_type

Type of bootstrap confidence intervals to calculate. (Default: "norm". Select "none" to avoid calculating bootstrap CIs.)

cell_size

Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)

level

Spatial level: 'cube', 'continent', 'country', 'world', 'sovereignty', or 'geounit'. (Default: 'cube')

region

The region of interest (e.g., "Europe"). (Default: "Europe")

ne_type

The type of Natural Earth data to download: 'countries', 'map_units', 'sovereignty', or 'tiny_countries'. (Default: "countries")

ne_scale

The scale of Natural Earth data to download: 'small' - 110m, 'medium' - 50m, or 'large' - 10m. (Default: "medium")

output_crs

The CRS you want for your calculated indicator. (Leave blank to let the function choose a default based on grid reference system)

first_year

Exclude data before this year. (Uses all data in the cube by default.)

last_year

Exclude data after this year. (Uses all data in the cube by default.)

spherical_geometry

If set to FALSE, will temporarily disable spherical geometry while the function runs. Should only be used to solve specific issues. (Default is TRUE)

make_valid

Calls st_make_valid() from the sf package. Increases processing time but may help if you are getting polygon errors. (Default is FALSE).

num_bootstrap

Set the number of bootstraps to calculate for generating confidence intervals. (Default: 1000)

crs_unit_convert

Force a particular output CRS even when it has different units than the input CRS. (Default: FALSE)

shapefile_path

Path of an external shapefile to merge into the workflow. For example, if you want to calculate your indicator particular features such as protected areas or wetlands.

invert

Calculate an indicator over the inverse of the shapefile (e.g. if you have a protected areas shapefile this would calculate an indicator over all non protected areas)

Value

An S3 object with the classes 'indicator_map' or 'indicator_ts' and 'pielou_evenness' or 'williams_evenness' containing the calculated indicator values and metadata.

Details

Evenness

Species evenness is a commonly used indicator that measures how uniformly individuals are distributed across species in a region or over time. It provides a complement to richness by taking relative abundance into account. Although GBIF provides information about abundances as individual counts, the majority of entries lack this information. Hence, evenness can only be calculated using the proportions of observations rather than proportions of individuals. Strictly speaking, the evenness measures therefore indicate how uniformly species are represented in the respective data set rather than the true evenness of the ecological community.

Pielou's evenness

Pielou's evenness (1966) is a well-known and commonly used evenness measure. It is calculated as:

$$ E = \frac{-\sum_{i=1}^{S} p_i \ln(p_i)}{\ln(S)} $$ where S is the number of species and pi is the proportion of occurrences represented by species i.

Williams' evenness

An analysis of evenness properties by Kvålseth (2015) showed that an evenness index introduced by Williams in 1977 in an unpublished manuscript has two important properties which Pielou's does not. The properties in question are complex mathematical properties known as the Schur-Concavity and value validity, but we attempt to describe them here more simply. If a measure of evenness is Schur-concave, it means that when the distribution of individuals becomes more evenly spread across species, the measure of evenness will stay the same or increase, but never decrease. Value validity means that an evenness index should provide sensible and meaningful values across its range for any given distribution of species abundances. Kvålseth referred to this evenness measure as E9 but we refer to it as Williams' evenness.

Williams' evenness is calculated as:

$$ 1 - \sqrt{\frac{S\sum_{i=1}^{S} p_i^2 - 1}{S - 1}} $$

where S is the number of species and pi is the proportion of occurrences represented by species i.

Functions

pielou_evenness_map():

pielou_evenness_ts():

williams_evenness_map():

williams_evenness_ts():

References

Pielou, E. C. (1966). The measurement of diversity in different types of biological collections. Journal of theoretical biology, 13, 131-144.

Kvålseth, T. O. (2015). Evenness indices once again: critical analysis of properties. SpringerPlus, 4, 1-12.

See also

Examples

if (FALSE) { # \dontrun{
pe_map <- pielou_evenness_map(example_cube_1, level = "country", region = "Denmark")
plot(pe_map)
} # }
if (FALSE) { # \dontrun{
pe_ts <- pielou_evenness_ts(example_cube_1, first_year = 1985)
plot(pe_ts)
} # }
if (FALSE) { # \dontrun{
we_map <- williams_evenness_map(example_cube_1, level = "country", region = "Denmark")
plot(we_map)
} # }
if (FALSE) { # \dontrun{
we_ts <- williams_evenness_ts(example_cube_1, first_year = 1985)
plot(we_ts)
} # }