Calculate Confidence Intervals for a Biodiversity Indicator

This function calculates bootstrap confidence intervals for a biodiversity indicator. It is called automatically when calculating a biodiversity indicator over time unless you choose 'none' for ci_type.

Usage

calc_ci(x, indicator, ...)

# S3 method for class 'total_occ'
calc_ci(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

# S3 method for class 'occ_density'
calc_ci(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

# S3 method for class 'newness'
calc_ci(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

# S3 method for class 'williams_evenness'
calc_ci(x, ...)

# S3 method for class 'pielou_evenness'
calc_ci(x, ...)

# S3 method for class 'ab_rarity'
calc_ci(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

# S3 method for class 'area_rarity'
calc_ci(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

# S3 method for class 'spec_occ'
calc_ci(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

# S3 method for class 'spec_range'
calc_ci(x, indicator, num_bootstrap = 1000, ci_type = ci_type, ...)

# S3 method for class 'tax_distinct'
calc_ci(
  x,
  indicator,
  num_bootstrap = 1000,
  ci_type = ci_type,
  set_rows = 1,
  ...
)

Arguments

x

A data cube object

indicator

An indicator calculated over time, in the form of a data frame. *Note: this should NOT be an 'indicator_ts' object as it is meant to be called by the 'compute_indicator_workflow' function.

...

Additional arguments passed to specific indicator calculation functions.

num_bootstrap

Set the number of bootstraps to calculate for generating confidence intervals. (Default: 1000)

ci_type

Type of bootstrap confidence intervals to calculate. (Default: "norm". Select "none" to avoid calculating bootstrap CIs.)

set_rows

Automatically select which taxonomic information to keep when there are multiple options. Default value of 1 keeps the first option, which is usually the best.