Calculate Biodiversity Indicators Over Space

This function provides a flexible framework for calculating various biodiversity indicators on a spatial grid or as a time series. It prepares the data, creates a grid, calculates indicators, and formats the output into an appropriate S3 object ('indicator_map'). Specific implementations for different indicator types are provided using the appropriate wrappers.

Usage

# S3 method for obs_richness
calc_map(x, ...)

# S3 method for total_occ
calc_map(x, ...)

# S3 method for newness
calc_map(x, newness_min_year = NULL, ...)

# S3 method for occ_density
calc_map(x, ...)

# S3 method for williams_evenness
calc_map(x, ...)

# S3 method for pielou_evenness
calc_map(x, ...)

# S3 method for ab_rarity
calc_map(x, ...)

# S3 method for area_rarity
calc_map(x, ...)

calc_map(x, ...)

Arguments

x

A data cube object ('processed_cube').

...

Additional arguments passed to specific indicator calculation functions.

type

The indicator to calculate. Supported options include:

'obs_richness': Observed species richness.

'total_occ': Total number of occurrences.

'newness': Mean year of occurrence.

'occ_density': Density of occurrences.

'williams_evenness', 'pielou_evenness': Evenness measures.

'ab_rarity', 'area_rarity': Abundance-based and area-based rarity scores.

cell_size

Length of grid cell sides, in km. (Default: 10 for country, 100 for continent or world)

level

Spatial level: 'continent', 'country', or 'world'. (Default: 'continent')

region

The region of interest (e.g., "Europe"). (Default: "Europe")

Value

An S3 object of the class 'indicator_map' containing the calculated indicator values and metadata.

Examples

observed_richness_map <- obs_richness_map(example_cube_2, level = "continent", region = "Europe")
#> Error in compute_indicator_workflow(data, type = "obs_richness", dim_type = "map",     ...): object 'x' not found
plot(observed_richness_map)
#> Error in eval(expr, envir, enclos): object 'observed_richness_map' not found