Calculate confidence intervals for list of boot objects

This function calculates confidence intervals for a list of objects of class "boot" per year into a dataframe containing all required summaries.

Usage

get_bootstrap_ci(bootstrap_list, ..., temporal_list_name = "year")

Arguments

bootstrap_list

A list of objects of class "boot" per year.

...

Additional argument to be passed to the boot::boot.ci() function.

temporal_list_name

The temporal list names of bootstrap_list (e.g., year, month ...) containing time point values. Default year.

Value

The returned value is a dataframe containing the time point, the type of interval (int_type), the lower limit of the confidence interval (ll), the upper limit of the confidence interval (ul), and the confidence level of the intervals (conf_level).