Extract Years With Observations from an Indicator Map

Takes an "indicator_map", "indicator_ts", or "processed_cube" object and determines the years for which observation data exists.

Usage

get_observed_years(x)

Arguments

x

An "indicator_map" object containing calculated indicator values associated with grid cells.

Value

A data frame with two columns:

years: A sequence of years covering the range of observations.

occurrences: A logical vector indicating if observations exist for each year (TRUE if present, FALSE if absent).

Examples

total_occ_insects_portugal <- total_occ_map(example_cube_2, level = "country", region = "Portugal")
get_observed_years(total_occ_insects_portugal)
#>     years occurrences
#> 1    1809        TRUE
#> 2    1810       FALSE
#> 3    1811       FALSE
#> 4    1812       FALSE
#> 5    1813       FALSE
#> 6    1814       FALSE
#> 7    1815       FALSE
#> 8    1816       FALSE
#> 9    1817       FALSE
#> 10   1818       FALSE
#> 11   1819       FALSE
#> 12   1820       FALSE
#> 13   1821       FALSE
#> 14   1822       FALSE
#> 15   1823       FALSE
#> 16   1824       FALSE
#> 17   1825       FALSE
#> 18   1826       FALSE
#> 19   1827       FALSE
#> 20   1828       FALSE
#> 21   1829       FALSE
#> 22   1830       FALSE
#> 23   1831       FALSE
#> 24   1832       FALSE
#> 25   1833       FALSE
#> 26   1834       FALSE
#> 27   1835       FALSE
#> 28   1836       FALSE
#> 29   1837       FALSE
#> 30   1838       FALSE
#> 31   1839       FALSE
#> 32   1840       FALSE
#> 33   1841       FALSE
#> 34   1842       FALSE
#> 35   1843       FALSE
#> 36   1844       FALSE
#> 37   1845       FALSE
#> 38   1846       FALSE
#> 39   1847       FALSE
#> 40   1848       FALSE
#> 41   1849        TRUE
#> 42   1850        TRUE
#> 43   1851       FALSE
#> 44   1852       FALSE
#> 45   1853       FALSE
#> 46   1854       FALSE
#> 47   1855       FALSE
#> 48   1856       FALSE
#> 49   1857       FALSE
#> 50   1858       FALSE
#> 51   1859       FALSE
#> 52   1860       FALSE
#> 53   1861       FALSE
#> 54   1862       FALSE
#> 55   1863       FALSE
#> 56   1864       FALSE
#> 57   1865       FALSE
#> 58   1866       FALSE
#> 59   1867       FALSE
#> 60   1868       FALSE
#> 61   1869       FALSE
#> 62   1870        TRUE
#> 63   1871       FALSE
#> 64   1872       FALSE
#> 65   1873        TRUE
#> 66   1874       FALSE
#> 67   1875       FALSE
#> 68   1876       FALSE
#> 69   1877       FALSE
#> 70   1878       FALSE
#> 71   1879       FALSE
#> 72   1880       FALSE
#> 73   1881        TRUE
#> 74   1882       FALSE
#> 75   1883       FALSE
#> 76   1884        TRUE
#> 77   1885        TRUE
#> 78   1886        TRUE
#> 79   1887        TRUE
#> 80   1888       FALSE
#> 81   1889       FALSE
#> 82   1890       FALSE
#> 83   1891       FALSE
#> 84   1892       FALSE
#> 85   1893        TRUE
#> 86   1894        TRUE
#> 87   1895        TRUE
#> 88   1896        TRUE
#> 89   1897        TRUE
#> 90   1898        TRUE
#> 91   1899        TRUE
#> 92   1900        TRUE
#> 93   1901        TRUE
#> 94   1902        TRUE
#> 95   1903        TRUE
#> 96   1904        TRUE
#> 97   1905       FALSE
#> 98   1906        TRUE
#> 99   1907        TRUE
#> 100  1908        TRUE
#> 101  1909        TRUE
#> 102  1910        TRUE
#> 103  1911        TRUE
#> 104  1912        TRUE
#> 105  1913        TRUE
#> 106  1914        TRUE
#> 107  1915        TRUE
#> 108  1916        TRUE
#> 109  1917       FALSE
#> 110  1918        TRUE
#> 111  1919       FALSE
#> 112  1920        TRUE
#> 113  1921        TRUE
#> 114  1922        TRUE
#> 115  1923        TRUE
#> 116  1924        TRUE
#> 117  1925        TRUE
#> 118  1926        TRUE
#> 119  1927        TRUE
#> 120  1928        TRUE
#> 121  1929        TRUE
#> 122  1930        TRUE
#> 123  1931        TRUE
#> 124  1932        TRUE
#> 125  1933        TRUE
#> 126  1934        TRUE
#> 127  1935        TRUE
#> 128  1936       FALSE
#> 129  1937        TRUE
#> 130  1938        TRUE
#> 131  1939        TRUE
#> 132  1940        TRUE
#> 133  1941        TRUE
#> 134  1942        TRUE
#> 135  1943        TRUE
#> 136  1944        TRUE
#> 137  1945        TRUE
#> 138  1946        TRUE
#> 139  1947        TRUE
#> 140  1948        TRUE
#> 141  1949        TRUE
#> 142  1950        TRUE
#> 143  1951        TRUE
#> 144  1952        TRUE
#> 145  1953        TRUE
#> 146  1954        TRUE
#> 147  1955        TRUE
#> 148  1956        TRUE
#> 149  1957        TRUE
#> 150  1958        TRUE
#> 151  1959        TRUE
#> 152  1960        TRUE
#> 153  1961        TRUE
#> 154  1962        TRUE
#> 155  1963        TRUE
#> 156  1964        TRUE
#> 157  1965        TRUE
#> 158  1966        TRUE
#> 159  1967        TRUE
#> 160  1968        TRUE
#> 161  1969        TRUE
#> 162  1970        TRUE
#> 163  1971        TRUE
#> 164  1972        TRUE
#> 165  1973        TRUE
#> 166  1974        TRUE
#> 167  1975        TRUE
#> 168  1976        TRUE
#> 169  1977        TRUE
#> 170  1978        TRUE
#> 171  1979        TRUE
#> 172  1980        TRUE
#> 173  1981        TRUE
#> 174  1982        TRUE
#> 175  1983        TRUE
#> 176  1984        TRUE
#> 177  1985        TRUE
#> 178  1986        TRUE
#> 179  1987        TRUE
#> 180  1988        TRUE
#> 181  1989        TRUE
#> 182  1990        TRUE
#> 183  1991        TRUE
#> 184  1992        TRUE
#> 185  1993        TRUE
#> 186  1994        TRUE
#> 187  1995        TRUE
#> 188  1996        TRUE
#> 189  1997        TRUE
#> 190  1998        TRUE
#> 191  1999        TRUE
#> 192  2000        TRUE
#> 193  2001        TRUE
#> 194  2002        TRUE
#> 195  2003        TRUE
#> 196  2004        TRUE
#> 197  2005        TRUE
#> 198  2006        TRUE
#> 199  2007        TRUE
#> 200  2008        TRUE
#> 201  2009        TRUE
#> 202  2010        TRUE
#> 203  2011        TRUE
#> 204  2012        TRUE
#> 205  2013        TRUE
#> 206  2014        TRUE
#> 207  2015        TRUE
#> 208  2016        TRUE
#> 209  2017        TRUE
#> 210  2018        TRUE
#> 211  2019        TRUE
#> 212  2020        TRUE