Extract Years With Observations from an Indicator Map

Takes an "indicator_map", "indicator_ts", or "processed_cube" object and determines the years for which observation data exists.

Usage

get_observed_years(x)

Arguments

x

An "indicator_map" object containing calculated indicator values associated with grid cells.

Value

A data frame with two columns:

years: A sequence of years covering the range of observations.

occurrences: A logical vector indicating if observations exist for each year (TRUE if present, FALSE if absent).

Examples

get_observed_years(example_indicator_map1)
#>    years occurrences
#> 1   1970        TRUE
#> 2   1971        TRUE
#> 3   1972        TRUE
#> 4   1973        TRUE
#> 5   1974        TRUE
#> 6   1975        TRUE
#> 7   1976        TRUE
#> 8   1977        TRUE
#> 9   1978        TRUE
#> 10  1979        TRUE
#> 11  1980        TRUE
#> 12  1981        TRUE
#> 13  1982        TRUE
#> 14  1983        TRUE
#> 15  1984        TRUE
#> 16  1985        TRUE
#> 17  1986        TRUE
#> 18  1987        TRUE
#> 19  1988        TRUE
#> 20  1989        TRUE
#> 21  1990        TRUE
#> 22  1991        TRUE
#> 23  1992        TRUE
#> 24  1993        TRUE
#> 25  1994        TRUE
#> 26  1995        TRUE
#> 27  1996        TRUE
#> 28  1997        TRUE
#> 29  1998        TRUE
#> 30  1999        TRUE
#> 31  2000        TRUE
#> 32  2001        TRUE
#> 33  2002        TRUE
#> 34  2003        TRUE
#> 35  2004        TRUE
#> 36  2005        TRUE
#> 37  2006        TRUE
#> 38  2007        TRUE
#> 39  2008        TRUE
#> 40  2009        TRUE
#> 41  2010        TRUE
#> 42  2011        TRUE
#> 43  2012        TRUE
#> 44  2013        TRUE
#> 45  2014        TRUE
#> 46  2015        TRUE
#> 47  2016        TRUE
#> 48  2017        TRUE
#> 49  2018        TRUE
#> 50  2019        TRUE
#> 51  2020        TRUE
#> 52  2021        TRUE
#> 53  2022        TRUE
#> 54  2023        TRUE