Process GBIF Data Cubes

Processes a GBIF data cube and associated taxonomic information file. All files must reside in the same directory and share a base file name (e.g., 'my_mammals_cube.csv', 'my_mammals_info.csv').

Usage

process_cube(
  cube_name,
  tax_info = NULL,
  datasets_info = NULL,
  first_year = NULL,
  last_year = NULL
)

Arguments

cube_name

The location and name of a data cube file (e.g., 'data/europe_species_cube.csv').

tax_info

The location and name of an associated taxonomic info file (e.g., 'data/europe_species_info.csv').

datasets_info

The location and name of an associated dataset info file (e.g., 'data/europe_species_datasets.csv').

first_year

(Optional) The first year of occurrences to include. If not specified, uses the earliest year present in the cube.

last_year

(Optional) The final year of occurrences to include. If not specified, uses the latest year present in the cube.

Value

A tibble containing the processed GBIF occurrence data.

Examples

europe_example_cube <- process_cube(cube_name = "data/europe_species_cube.csv",
                                    tax_info = "data/europe_species_info.csv")
#> Error: 'data/europe_species_cube.csv' does not exist in current working directory ('/home/runner/work/b3gbi/b3gbi/docs/reference').
europe_example_cube
#> Error in eval(expr, envir, enclos): object 'europe_example_cube' not found